Refatoração geral
Adicionado comentários
Trocando nome de métodos,
Otimizações no calculo do fitness
Melhorando Makefile (agora ele compila com FFLAGS)

+++ mapa/main.c
@@ -9,8 +9,8 @@ static const int N = 19683;   // powi(3, 9); numero de arranjos possiveis de X e
 static const int M = 986410;  //409114;  // 1 + 9 + 9*8 + 9*8*7 + ... + 9*8*7*6*5*4*3*2*1 ; Numemro maximo de recursoes de jogo_possiveis()
 
 // extern int simetrias[8][9];
-extern int simetrias_reversa[8][9];
-extern int vitorias[8][3];
+extern byte simetrias_reversa[8][9];
+extern byte vitorias[8][3];
 
 /**
  * Salva mapa em uma arquivo em path
@@ -48,7 +48,7 @@ falha:
  * com todas as jogadas a ser feita seja menor
  */
 int main() {
-    int** jogos = NULL;
+    byte** jogos = NULL;
 
     int i;
     bool ok;
