+++ minmax/minimax.h
@@ -5,18 +5,18 @@
 #include <stdlib.h>
 #include "util.h"
 
-extern int vitorias[8][3];
+extern byte vitorias[8][3];
 
 /**
- * Retorna se aquela jogada resultara em uma vitoria, empate ou derrota
+ * Retorna se aquela jogada resultara em uma vitoria (1), empate (0) ou derrota (-1)
  */
-int minimax(int* jogo, int vez, int maximizador);
+int minimax(byte* jogo, byte vez, byte maximizador);
 
 /**
  * Retorna entre as jogadas possiveis uma que resultara em vitoria
  * Senao pelo menos uma que resultara em empate
  * Senao retorna uma entre as jogadas disponiveis
  */
-int melhor_jogada(int* jogo);
+int melhor_jogada(byte* jogo);
 
 #endif
